badplace

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: mvautrot <mvautrot@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/12 11:25:31 by mvautrot          #+#    #+#             */
-/*   Updated: 2022/12/27 10:40:49 by mvautrot         ###   ########.fr       */
+/*   Updated: 2022/12/28 10:34:52 by mvautrot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -99,7 +99,7 @@ int		ft_hexa_min(unsigned int nb);
 int		ft_hexa_max(unsigned int nb);
 int		ft_putaddr(uintptr_t nb);
 int		ft_hexa_ptr(unsigned long long nb);
-size_t	ft_strlen( char *str);
+int	ft_strlen( char *str);
 
 
 #endif
